+++ .gitignore
@@ -1,9 +1,10 @@
+.DS_Store
 node_modules
-
+/build
+/.svelte-kit
+/package
 .env
-
-package-lock.json
-
-tests
-
-old
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
