+++ .npmrc
@@ -1,1 +1,1 @@
-@littlepriceonu:registry=https://npm.pkg.github.com
+engine-strict=true
