--- build.bat
@@ -1,4 +1,0 @@
-@echo off
-
-start npx tailwindcss -i ./src/styles.css -o ./dist/output.css --watch --minify
-start tsc -w
