--- nginx.conf
@@ -1,15 +1,0 @@
-events {
-
-}
-
-http {
-  server {
-    listen 80;
-    server_name littlepriceonu.com www.littlepriceonu.com;
-    location / {
-      proxy_set_header X-Real-IP $remote_addr;
-      proxy_set_header Host $http_host;
-      proxy_pass http://127.0.0.1:3000;
-    }
-  }
-}
