+++ postcss.config.js
@@ -1,0 +1,8 @@
+export default {
+    plugins: {
+        "postcss-import": {},
+        "tailwindcss/nesting": {},
+        "tailwindcss": {},
+        "autoprefixer": {}
+    },
+}
