--- src/404.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <!-- ! WARNING WARNING WARNING ! -->
-    <!-- USER SEEMS TO BE BROWSING SOURCE. CONTINUE TO MONITOR FOR FURTHER IMPROPER USE -->
-    <!-- ! MONITORING HAS BEEN ESCALATED TO A LEVEL *1* ! -->
-
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <link rel="canonical" href="https://littlepriceonu.com/404">
-    <meta name="robots" content="index">
-    <link rel="shortlink" href="https://littlepriceonu.com/404">
-    <meta name="description" content="Littlepriceonu | DATA TRANSFER ERROR">
-
-    <meta property="og:site_name" content="DATA LOST">
-    <meta property="og:type" content="website">
-    <meta property="og:url" content="https://littlepriceonu.com/404">
-    <meta property="og:title" content="How did you even get here?">
-    <meta property="og:description" content="FOREIGN CONNECTION DETECTED.">
-    <meta property="og:image" content="https://littlepriceonu.com/PFP.gif">
-    <meta property="og:image:alt" content="Too Late... They're in...">
-
-    <meta name="twitter:description" content="SAFETY DISABLED">
-
-    <meta name="MobileOptimized" content="width">
-    <meta name="HandheldFriendly" content="true">
-
-    <title>Littlepriceonu | How did you even get here?</title>
-
-    <link rel="stylesheet" href="/index.css">
-</head>
-
-<body class="w-full h-full bg-slate-900 scrollbar scrollbar-track-slate-900 scrollbar-thumb-stone-900 pl-6 pr-6 xl:pl-14 xl:pr-14 overflow-y-hidden">
-    <div class="w-full h-full bg-transparent flex flex-col items-center justify-center">
-        <h1 class="text-5xl text-blue-700 font-bold drop-shadow-[0_0_15px_blue]">404: Not Found</h1>
-        <h2 class="pt-4">How did you even get <a class="text-blue-400 hover:underline font-mono" href="/">here</a>?</h2>
-    </div>
-</body>
-
-</html>
