--- src/index.html
@@ -1,117 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="fixed">
-
-<head>
-    <!-- ! WARNING WARNING WARNING ! -->
-    <!-- USER SEEMS TO BE BROWSING SOURCE. CONTINUE TO MONITOR FOR FURTHER IMPROPER USE -->
-    <!-- ! MONITORING HAS BEEN ESCALATED TO A LEVEL *1* ! -->
-
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <link rel="canonical" href="https://littlepriceonu.com/">
-    <meta name="robots" content="index">
-    <link rel="shortlink" href="https://littlepriceonu.com/">
-    <meta name="description" content="Littlepriceonu | ACCESS GRANTED">
-
-    <meta property="og:site_name" content="DATA TRANSFER COMPLETE">
-    <meta property="og:type" content="website">
-    <meta property="og:url" content="https://littlepriceonu.com/">
-    <meta property="og:title" content="ACCESS GRANTED">
-    <meta property="og:description" content="FOREIGN CONNECTION DETECTED.">
-    <meta property="og:image" content="https://littlepriceonu.com/PFP.gif">
-    <meta property="og:image:alt" content="Too Late... They're in...">
-
-    <meta name="twitter:description" content="Did you SERIOUSLY not terminate the connection?">
-
-    <meta name="MobileOptimized" content="width">
-    <meta name="HandheldFriendly" content="true">
-
-    <title>Littlepriceonu | ACCESS GRANTED</title>
-
-    <link rel="stylesheet" href="index.css">
-    <script type="module" src="scripts/index.js" defer></script>
-    
-    <script src="https://unpkg.com/node-vibrant@3.1.5/dist/vibrant.min.js"></script>
-</head>
-
-<body class="w-full h-full bg-slate-900 scrollbar scrollbar-track-slate-900 scrollbar-thumb-stone-900 pl-6 pr-6 xl:pl-14 xl:pr-14 overflow-y-hidden">
-    <div id="loadingScreen" class="flex flex-col justify-center items-center absolute z-20 top-0 left-0 w-full h-full">
-        <h2 id="lennyFace" class="text-xl text-white text-center w-96" ></h2>
-
-        <h1 class="text-4xl font-mono text-center pb-2 pt-1">! Loading !</h1>
-
-        <div id="loadingBarBackground" class="h-6 w-96 bg-slate-700 rounded-md">
-            <div id="loadingBar" class="h-full w-0 z-30 rounded-md"></div>
-        </div>
-    </div>
-
-    <div id="contentHolder" class="w-full h-full bg-slate-900 opacity-0">
-        <header class="sticky top-0 bg-transparent w-full h-headerHeight flex flex-row items-center justify-end space-x-12">
-            <h1 id="accessGranted" class="hidden xl:block font-mono mr-auto text-xl text-green-500 font-semibold cursor-pointer select-none" onclick="console.log('%cUSER SEEMS TO HAVE INTERACTED WITH THE TERMINAL... HOW... ' + '%cSTRANGE' + '%c...', 'color: red; font-family: monospace;', 'font-family: monospace; font-weight: 800; color: green;', 'font-family: monospace; color: red;')" >ACCESS GRANTED</h1>
-            
-            <a class="text-3xl font-bold text-transparent bg-clip-text bg-gradientSize bg-gradient-to-r from-cyan-500 to-blue-900 animate-fastScrollingGradient motion-reduce:animate-none motion-reduce:bg-auto" href="/">Home</a>
-            <a class="text-3xl" href="">More Soon ™️</a>
-            <!-- <a class="text-3xl" href="">Projects</a>
-            <a class="text-3xl" href="">Something Else</a> -->
-        </header>
-    
-        <div id="MainContent" class="h-full pt-[4.5rem] pl-0 xl:pl-24">
-            <img id="PFP" class="h-44 rounded-lg relative top-[40px]" src="/img/PFP.gif" alt="Vault Boy Being Tired :>">
-            
-            <div id="OuterStatus" class="rounded-full h-12 w-12 bg-slate-900 z-10 relative flex items-center justify-center">
-                    <div id="Status" data-status="offline" class="transition-all duration-300 rounded-full w-6 h-6 bg-green-500">
-                        <div id="dndBar" class="w-4 h-[4px] bg-red-900 class hidden opacity-0 transition-[opacity] duration-700"></div>
-                        <div id="idleCircle" class="w-6 h-6 rounded-full bg-slate-900 class hidden opacity-0 transition-[opacity] duration-1000"></div>
-                    </div>
-            </div>
-
-            <h1 class="text-5xl font-bold">Littlepriceonu</h1>
-            <h2 class="text-2xl pt-1">Full Stack <mark id="developerText" class="font-bold text-transparent bg-miniGradientSize animate-straightScrollingGradient bg-clip-text bg-gradient-to-r from-blue-600 to-indigo-900 motion-reduce:animate-none motion-reduce:bg-auto">Web Developer</mark></h2>
-    
-            <div id="Socials" class="flex flex-row items-center justify-start space-x-6 pt-5">
-                <a target="_blank" href="https://github.com/Littlepriceonu">
-                    <img class="h-12 select-none" src="/img/github.png" alt="Github Logo">
-                </a>
-    
-                <a target="_blank" href="https://discord.com/users/526120594929090561">
-                    <img class="h-12 select-none" src="/img/discord.png" alt="Discord Logo">
-                </a>
-    
-                <a target="_blank" href="https://steamcommunity.com/id/littlepriceonu/">
-                    <img class="h-12 select-none" src="/img/steam.png" alt="Steam Logo">
-                </a>
-    
-                <a target="_blank" href="https://twitter.com/littlepriceonu">
-                    <img class="h-12 select-none" src="/img/twitter.png" alt="Steam Logo">
-                </a>
-    
-                <a target="_blank" href="https://www.last.fm/user/littlepriceonu">
-                    <img class="h-12 select-none" src="/img/lastfm.png" alt="Steam Logo">
-                </a>
-            </div>
-    
-            <div id="ListeningTo" class="relative z-50 flex flex-col items-center justify-start mt-5 3xl:mt-10 w-fit min-w-[85%] md:min-w-[60%] lg:max-w-[45%] lg:min-w-[40.333333%] 2xl:max-w-[30%] 2xl:min-w-[25%] h-fit min-h-[28%] 2xl:min-h-[26%]  bg-gray-800 rounded-md">
-                <div class="w-full flex flex-row justify-start items-center"> <h1 id="NowListening" class="font-bold pt-2 w-full text-left pl-3 font-mono" >Loading Module...</h1> 
-                    <div class="bar w-[3px] h-6 bg-blue-600 mr-1 mt-2"></div>
-                    <div class="bar w-[3px] h-3 bg-blue-600 mr-1 mt-2"></div>
-                    <div class="bar w-[3px] h-5 bg-blue-600 mr-2 mt-2"></div>
-                </div>
-                <div class="flex w-full flex-row mt-4 justify-start items-center pl-6 relative">
-                    <img id="BlurImage" crossorigin="anonymous" class="top-0 mb-auto h-[8.5rem] blur-lg rounded-md absolute" src="" alt="Loading Image...">
-                    <img id="AlbumImage" crossorigin="anonymous" class="ml-1.5 mb-auto h-32 rounded-md z-10" src="" alt="Loading Image...">
-                    <div class="w-full flex flex-col items-center justify-center ml text-xl pl-5 text-center">
-                        <h1 id="SongName" class="pr-5 font-bold"> Loading... </h1>
-                        <p  id="AlbumName" class="text-sm pt-1 text-gray-400 pr-5">Loading...</p> 
-                        <p  id="ArtistName" class="text-sm pt-3 text-gray-400 pb-3 pr-5">Loading...</p>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <div class="absolute top-0 left-0 pt-"></div>
-    </div>
-</body>
-
-</html>
