+++ src/routes/+layout.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+    import "../app.css"
+</script>
+
+<svelte:head>
+    <meta property="og:site_name" content="DATA TRANSFER COMPLETE | LITTLEPRICEONU">
+    <meta property="og:type" content="website">
+    <meta property="og:url" content="https://littlepriceonu.com/">
+</svelte:head>
+
+<slot />
